exercise 4.21 solved

+++ part4/bloglist/controllers/blogs.js
@@ -70,7 +70,29 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if (!user) {
+    return response.status(400).json({ error: 'user not found' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).end()
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'unauthorized' })
+  }
+  
+  user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+  await user.save()
   await Blog.findByIdAndDelete(request.params.id)
+
   response.status(204).end()
 })
